clean up form
